requirements.txt
websockets>=10.0
pymongo>=4.0
python-dotenv 